fix: Enable SFTP
sshd_config to enable SFTP subsystem

Recent build-agent (picard containers on Nomad clients) has OpenSSH >= 9.0, with which scp command is no longer based on SCP, and is on SFTP instead. This PR follows that move and enables SFTP subsystem on ephemeral machines for Windows builds by default.
https://www.openssh.com/txt/release-9.0

Originally filed by Makoto but for a lot of reasons I had to file this on his behalf:
https://github.com/CircleCI-Public/circleci-server-windows-image-builder/pull/15

(No Jira)